Updated position field type for Section model of instruction app

+++ instruction/models.py
@@ -11,7 +11,7 @@ def section_position_default() -> int:
 class Section(models.Model): # type: ignore [django-manager-missing]
 	title = models.CharField(_('Заголовок'), max_length=255)
 	text = models.TextField(_('Текст'))
-	position = models.IntegerField(_('Позиция'), blank=True, default=section_position_default)
+	position = models.PositiveSmallIntegerField(_('Позиция'), blank=True, default=section_position_default)
 
 	class Meta(TypedModelMeta):
 		ordering = ('position',)
